@@ -9,7 +9,7 @@ readme = "README.md"
 [tool.poetry.dependencies]
 channels = "^4.0.0"
 crispy-bootstrap5 = "^2024.2"
-cryptography = "^42.0.5"
+cryptography = "^43.0.0"
 daphne = "^4.0.0"
 Django = "^5.0.4"
 django-crispy-forms = "^2.0"
@@ -24,11 +24,11 @@ django-tables2 = "^2.3.3"
 djangorestframework = "^3.13.1"
 procrastinate = { extras = ["django"], version = "^2.8.0" }
 psycopg = { extras = ["binary"], version = "^3.1.12" }
-python = "^3.11"
+python = ">=3.12,<4.0"
 toml = "^0.10.2"
 Twisted = { extras = ["tls", "http2"], version = "^24.3.0" }
 wait-for-it = "^2.2.2"
-watchfiles = "^0.22.0"
+watchfiles = "^0.24.0"
 whitenoise = "^6.0.0"
 
 [tool.poetry.group.dev.dependencies]
@@ -40,13 +40,13 @@ django-test-migrations = "^1.3.0"
 djangorestframework-stubs = "^3.15.0"
 djlint = "^1.19.16"
 factory-boy = "^3.1.0"
-Faker = "^25.4.0"
+Faker = "^28.1.0"
 invoke = "^2.1.2"
 ipykernel = "^6.17.1"
 ipython = "^8.1.1"
 pyright = "^1.1.351"
 pytest = "^8.1.1"
-pytest-asyncio = "^0.23.6"
+pytest-asyncio = "^0.24.0"
 pytest-cov = "^5.0.0"
 pytest-django = "^4.5.2"
 pytest-mock = "^3.10.0"
@@ -57,7 +57,7 @@ pytest-timeout = "^2.1.0"
 pytest-watch = "^4.2.0"
 python-dotenv = "^1.0.0"
 requests = "^2.27.1"
-ruff = "^0.5.1"
+ruff = "^0.6.3"
 time-machine = "^2.1.0"
 django-browser-reload = "^1.11.0"
 pywatchman = "^2.0.0"
